+++ First/gotoeat/gotoeat-api/src/main/java/kr/co/first/gotoeat/domain/RestaurantRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RestaurantRepository extends CrudRepository<Restaurant, Integer>
+public interface RestaurantRepository extends CrudRepository<Restaurant, Long>
 {
     List<Restaurant> findAll();
